requirements.txt
#
# This file is autogenerated by pip-compile with python 3.9
# To update, run:
#
#    pip-compile
#
anyio==3.6.1
    # via starlette
asgiref==3.5.2
    # via uvicorn
attrs==22.1.0
    # via pytest
certifi==2022.6.15
    # via
    #   minio
    #   requests
chardet==4.0.0
    # via requests
click==8.1.3
    # via uvicorn
dbutils==2.0.2
    # via -r requirements.in
exceptiongroup==1.0.4
    # via pytest
fastapi==0.78.0
    # via
    #   -r requirements.in
    #   fastapi-utils
fastapi-utils==0.2.1
    # via -r requirements.in
filetype==1.0.7
    # via -r requirements.in
h11==0.13.0
    # via uvicorn
idna==2.10
    # via
    #   anyio
    #   requests
iniconfig==1.1.1
    # via pytest
markupsafe==2.1.1
    # via werkzeug
marshmallow==3.14.1
    # via
    #   -r requirements.in
    #   webargs
minio==7.1.2
    # via -r requirements.in
mypy-extensions==0.4.3
    # via typing-inspect
numpy==1.20.3
    # via
    #   -r requirements.in
    #   opencv-python
    #   pandas
    #   pandera
    #   pyarrow
opencv-python==4.5.5.64
    # via -r requirements.in
packaging==21.3
    # via
    #   pandera
    #   pytest
pandas==1.2.4
    # via
    #   -r requirements.in
    #   pandera
pandas-stubs==1.2.0.61
    # via pandera
pandera==0.8.0
    # via -r requirements.in
pillow==9.0.0
    # via -r requirements.in
pluggy==1.0.0
    # via pytest
pyarrow==8.0.0
    # via pandera
pydantic==1.9.1
    # via
    #   -r requirements.in
    #   fastapi
    #   fastapi-utils
pyjwt==2.1.0
    # via -r requirements.in
pymongo==4.2.0
    # via -r requirements.in
pyodbc==4.0.30
    # via -r requirements.in
pyparsing==3.0.9
    # via packaging
pytest==7.2.0
    # via pytest-env
pytest-env==0.6.2
    # via -r requirements.in
python-dateutil==2.8.1
    # via
    #   -r requirements.in
    #   pandas
python-dotenv==0.17.1
    # via -r requirements.in
python-multipart==0.0.5
    # via -r requirements.in
pytz==2022.1
    # via pandas
requests==2.25.1
    # via -r requirements.in
six==1.16.0
    # via
    #   python-dateutil
    #   python-multipart
sniffio==1.2.0
    # via anyio
sqlalchemy==1.3.18
    # via
    #   -r requirements.in
    #   fastapi-utils
starlette==0.19.1
    # via fastapi
tomli==2.0.1
    # via pytest
typing-extensions==4.2.0
    # via
    #   pydantic
    #   starlette
    #   typing-inspect
typing-inspect==0.7.1
    # via pandera
urllib3==1.25.9
    # via
    #   -r requirements.in
    #   minio
    #   requests
uvicorn==0.17.6
    # via -r requirements.in
webargs==8.0.1
    # via -r requirements.in
werkzeug==2.2.2
    # via -r requirements.in
wrapt==1.14.1
    # via pandera
